Fix call to fcntl() with F_GETFL
Looks like the call to this function was done incorrectly according
to the fcntl() documentation I have been reading. This call returns
the flags set on the descriptor and the argument is actually ignored.
The worst part of all of this is that there was no error checking so
the call to put a socket into non-blocking mode would sometimes work
and sometimes not work based solely on what you might find randomly in
memory.