Convert ipaddr in conn_entry struct from pointer to char array
I was stupid to think that this would work before. Previously we were just
assigning a pointer to a pointer. Each connection and getpeername() call
would rewrite the address in memory which our pointer looked at. Instead
copy the contents of that memory in our char array and store it so this
cannot happen.