Fix issues with packet age comparison
I knew checking compiler warnings was a good idea. Some use of uninitialized
variable warnings lead me to find that the packet age comparison seems
to have been broken since da7767c4.

While the idea for the fix was good and I agree with the intended change
it seems as those the code that actually initializes the time was removed
for some reason.

The other interesting part is that this seems to not break for the most part
which is probably why it was accepted in the first place and why no one has really
complained about it.

Seems as though whatever is set in the stack is done so consistantly enough that
the packet age comparison would never break.

I'm told by people much smarter than me that this isn't very uncommon and is a good
enough explaination for me as to how something this broken could have continued to
work consistantly for this long without people noticing.