Fixes to smaller packet size logic
It seems that in 3950d852 some logic was added to make it so older clients
that use a smaller packet size could still communicate with newer servers.
While I think this is a good thing I'm not sure that the logic was added
in the right place. It was getting done inside an if statement used to
handle issues with recv() which doesn't really make sense to me since
after doing this the handle_read_events() would just return.

Will need to do some testing with older clients to see if this breaks
and to see if this ever worked.