Fix potential crash due to libwrap functionality
Found that when using libwrap there is the potential for NSCA to crash
when trying to determine if the connecting host should be allow to
communicate with NSCA. The problem stemmed from only passing the file
descriptor when initializing the request_info struct. Then a call to
fromhost() would try to determine the client/server IP addresses using
sock_host() and getpeername().

This is all well and good most of the time but the other day NSCA
crashed because the call to getpeername() could not be completed due to
the connection getting reset by peer. Here is the full error message:

nsca[15958]: warning: can't get client address: Connection reset by peer

My guess is after the getpeername() call was complete, fromhost()
returned without ever setting the client/server address information (as
it should) but calling hosts_access() without that information lead to
some type of crash.

Unfortunately, there wasn't any data about the crash in messages so it
is hard to say what all went wrong. This is more of an educated guess at
this point but I have a strong feeling I'm on the right track.

The solution (or so I hope) is to eliminate the call to fromhost() and
when initializing the request_info struct pass in the sockaddr_in
structs for both the client and server. That in addition to a call to
sock_methods() means the libwrap library should have all it needs to get
address information without any calls to things like getpeername(). This
should hopefully eliminate the failure scenario discussed previously.

This new method was mostly modeled after how socat implements libwrap.
Also, the FreeBSD source code on github was very helpful in figuring out
how all of the libwrap stuff worked so much thanks to them.

One thing to note is that this new method returns the server data
differently than the old. For example, when running the NSCA server
binding to any address, the previous method would return '127.0.0.1'
while the new method returns '0.0.0.0'. I'm not sure why the difference
yet nor do I care all that much. It does mean if people were using the
old method and creating denies like 'nsca@127.0.0.1: 192.168.1.10'
than they will have to switch. I imagine that people mostly just do
'nsca: 192.168.1.10'.