Make libwrap support optional
Make using libwrap an option when running configure. Not everyone needs
this so making it an option is best. Disabling libwrap can cut down on
complexity, processing necessary per connection and potential bugs in
the libwrap code. It is still enabled by default so this should not
break any existing configure strings.